Adopt main's MActor member and accessor naming: mMActor->unk4 becomes
mMActor->mModel, and the getUnkC() rename this branch carried is dropped in
favour of main's getAnmBck(), which leaves gatekeeper.cpp, hinokuri2.cpp and
NpcInitPrg.cpp identical to main again.

MActor::getCurBckAnmPtr() is kept alongside getAnmBck(). It is a real retail
symbol -- mario.MAP lists getCurBckAnmPtr__6MActorFv (UNUSED, 0x1c) for
bosstelesa.cpp -- so removing it drops a symbol the map requires.

+++ include/Player/MarioCap.hpp
@@ -18,7 +18,7 @@ public:
 
 	TMarioCap(TMario*);
 
-	virtual void perform(unsigned long cue, JDrama::TGraphics* graphics);
+	virtual void perform(u32 cue, JDrama::TGraphics* graphics);
 
 	void createMirrorCap();
 	void mtxEffectHide();
